requirements.txt
streamlit>=1.32.0
google-genai>=0.3.0
plotly>=5.18.0
python-dotenv>=1.0.0
Pillow>=10.0.0
streamlit
google-generativeai>=0.8.3
pandas
plotly